requirements.txt
Flask==3.0.2
GPUtil==1.4.0
gradio==3.41.2
gradio_client==0.5.0
Pillow==10.2.0
psutil==5.9.5
py_cpuinfo==9.0.0
